Start of Recipe page
Includes header with publisher username, date, meal type, and cuisine type, along
with a list of ingredients. Since we only have ingredient lists for
Banana Bread and Banana Pudding, these are the only pages that display
right now until we populate the ingredients_list table with more entries.

Changed routing in recipe_card to the page for the specific recipe_name.